requirements.txt
django
Pillow
gunicorn
whitenoise
